DEV: Autoload plugin files and namespace some jobs
